Moves (deprecated) SearchFilterDrawer to ol-components
