+++ frontends/ol-components/src/components/SearchFilterDrawer/FacetDisplay.test.tsx
