+++ frontends/ol-components/src/index.ts
@@ -103,6 +103,7 @@ export * from "./components/ChipLink/ChipLink"
 export * from "./components/FormDialog/FormDialog"
 export * from "./components/RoutedDrawer/RoutedDrawer"
 export * from "./components/SearchInput/SearchInput"
+export * from "./components/SearchFilter/SearchFilter"
 export * from "./components/SimpleMenu/SimpleMenu"
 export * from "./components/ShareTooltip/ShareTooltip"
 export * from "./components/TitledCarousel/TitledCarousel"
