+++ frontends/ol-search-ui/src/components/index.ts
@@ -4,5 +4,3 @@ export type {
   OnActivateCard,
   CardVariant,
 } from "./LearningResourceCardTemplate"
-
-export { default as SearchFilterDrawer } from "./SearchFilterDrawer"
